backend/src/covenant/api/ultimate_api.py
"""
Ultimate FastAPI Application with GraphQL, gRPC, WebSocket, and Advanced Features
"""
from fastapi import FastAPI, WebSocket, Depends, HTTPException, BackgroundTasks
from fastapi.middleware.cors import CORSMiddleware
from fastapi.middleware.gzip import GZipMiddleware
from fastapi.middleware.trustedhost import TrustedHostMiddleware
import strawberry
from strawberry.fastapi import GraphQLRouter
from prometheus_client import make_asgi_app
import uvicorn
from typing import List, Optional
import asyncio

from covenant.core.ultimate_engine import create_ultimate_engine, VerificationLevel
from covenant.config import settings

# GraphQL Schema
@strawberry.type
class EvaluationResult:
    action_id: str
    is_allowed: bool
    confidence: float
    execution_time_ms: float

@strawberry.type
class Query:
    @strawberry.field
    async def health(self) -> str:
        return "healthy"
    
    @strawberry.field
    async def metrics(self) -> str:
        engine = create_ultimate_engine()
        return str(engine.get_comprehensive_metrics())

@strawberry.type
class Mutation:
    @strawberry.field
    async def evaluate_action(self, action_data: str) -> EvaluationResult:
        engine = create_ultimate_engine()
        result = await engine.evaluate_action_ultimate({"data": action_data})
        return EvaluationResult(
            action_id=result["action_id"],
            is_allowed=result["is_allowed"],
            confidence=result["confidence"],
            execution_time_ms=result["execution_time_ms"]
        )

schema = strawberry.Schema(query=Query, mutation=Mutation)

# FastAPI App
app = FastAPI(
    title="COVENANT.AI Ultimate v4.0",
    version="4.0.0",
    description="Most Advanced Constitutional AI Platform Ever Built"
)

# Middleware
app.add_middleware(CORSMiddleware, allow_origins=["*"], allow_credentials=True, 
                   allow_methods=["*"], allow_headers=["*"])
app.add_middleware(GZipMiddleware, minimum_size=1000)

# GraphQL endpoint
graphql_app = GraphQLRouter(schema)
app.include_router(graphql_app, prefix="/graphql")

# Metrics
metrics_app = make_asgi_app()
app.mount("/metrics", metrics_app)

# WebSocket endpoint
@app.websocket("/ws/evaluate")
async def websocket_evaluate(websocket: WebSocket):
    await websocket.accept()
    engine = create_ultimate_engine()
    while True:
        data = await websocket.receive_json()
        result = await engine.evaluate_action_ultimate(data)
        await websocket.send_json(result)

# REST API
@app.get("/")
async def root():
    return {
        "name": "COVENANT.AI Ultimate",
        "version": "4.0.0",
        "features": [
            "Multi-agent swarm coordination",
            "Quantum-inspired optimization",
            "Zero-knowledge proofs",
            "Differential privacy",
            "Homomorphic encryption",
            "Meta-learning",
            "Federated learning",
            "Certified robustness",
            "Formal verification",
            "Causal inference",
            "GraphQL API",
            "WebSocket streaming",
            "gRPC support"
        ]
    }

@app.post("/api/v4/evaluate")
async def evaluate_ultimate(
    action: dict,
    verification_level: str = "FORMAL",
    generate_proof: bool = True
):
    engine = create_ultimate_engine()
    level = VerificationLevel[verification_level]
    return await engine.evaluate_action_ultimate(
        action, verification_level=level, generate_proof=generate_proof
    )

if __name__ == "__main__":
    uvicorn.run(app, host="0.0.0.0", port=8000)
