backend/requirements-full.txt
# Core Framework (Latest Stable)
fastapi[all]==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
starlette==0.36.3

# ASGI Servers & Performance
uvloop==0.19.0
httptools==0.6.1

# Database Layer
sqlalchemy[asyncio]==2.0.27
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
databases[postgresql]==0.8.0

# Advanced ORM
sqlmodel==0.0.16
tortoise-orm==0.20.0

# Caching & Message Queue
redis[hiredis]==5.0.1
aioredis==2.0.1
celery[redis]==5.3.6
dramatiq[redis]==1.15.0
rq==1.16.1

# Task Scheduling
apscheduler==3.10.4
croniter==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt,argon2]==1.7.4
PyJWT==2.8.0
argon2-cffi==23.1.0
bcrypt==4.1.2
cryptography==42.0.2
python-multipart==0.0.7
itsdangerous==2.1.2

# OAuth & SSO
authlib==1.3.0
python-social-auth==0.3.6

# Monitoring & Observability
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-sqlalchemy==0.43b0
opentelemetry-instrumentation-redis==0.43b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-prometheus==0.43b0
sentry-sdk[fastapi]==1.40.3
structlog==24.1.0
python-json-logger==2.0.7

# AI/ML Frameworks
torch==2.2.0
torchvision==0.17.0
transformers==4.37.2
sentence-transformers==2.3.1
numpy==1.26.4
scipy==1.12.0
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0
catboost==1.2.2

# Deep Learning
tensorflow==2.15.0
jax==0.4.23
flax==0.8.0

# Explainability & Fairness
shap==0.44.1
lime==0.2.0.1
fairlearn==0.10.0
aif360==0.6.0

# NLP & Embeddings
spacy==3.7.3
nltk==3.8.1
gensim==4.3.2
fasttext==0.9.2

# Formal Verification
z3-solver==4.12.5.0
sympy==1.12
networkx==3.2.1
pulp==2.8.0

# Blockchain & Crypto
web3==6.15.1
eth-account==0.11.0
eth-utils==4.0.0
py-evm==0.8.0b1
ecdsa==0.18.0
pycryptodome==3.20.0

# Graph Processing
neo4j==5.17.0
py2neo==2021.2.3

# Vector Databases
pinecone-client==3.0.2
chromadb==0.4.22
weaviate-client==4.4.2
qdrant-client==1.7.3
milvus==2.3.5

# Time Series
influxdb-client==1.39.0
timescaledb==0.0.4

# Search
elasticsearch==8.12.0
opensearch-py==2.4.2

# GraphQL & gRPC
strawberry-graphql[fastapi]==0.219.2
graphene==3.3
ariadne==0.22.0
grpcio==1.60.1
grpcio-tools==1.60.1
protobuf==4.25.2

# HTTP & WebSocket
httpx==0.26.0
aiohttp==3.9.3
websockets==12.0
httpcore==1.0.2
requests==2.31.0

# Data Validation & Serialization
marshmallow==3.20.2
cerberus==1.3.5
jsonschema==4.21.1
orjson==3.9.13
msgpack==1.0.7
pyarrow==15.0.0

# Data Processing
polars==0.20.7
pandas==2.2.0
dask==2024.2.0
ray==2.9.1

# Image & Media
pillow==10.2.0
opencv-python==4.9.0.80
imageio==2.33.1

# PDF & Documents
pypdf==4.0.1
reportlab==4.1.0
weasyprint==60.2
python-docx==1.1.0
openpyxl==3.1.2

# Cloud Services
boto3==1.34.44
google-cloud-storage==2.14.0
azure-storage-blob==12.19.0
minio==7.2.3

# Feature Flags & Config
launchdarkly-server-sdk==9.2.1
python-decouple==3.8
environs==11.0.0
dynaconf==3.2.4

# Rate Limiting
slowapi==0.1.9
pyrate-limiter==3.1.1

# Caching Strategies
aiocache==0.12.2
cachetools==5.3.2
diskcache==5.6.3

# API Documentation
mkdocs==1.5.3
mkdocs-material==9.5.8
mkdocstrings[python]==0.24.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
hypothesis==6.98.1
faker==22.6.0
factory-boy==3.3.0
locust==2.20.1
pytest-benchmark==4.0.0

# Code Quality
black==24.1.1
ruff==0.2.1
mypy==1.8.0
pylint==3.0.3
bandit==1.7.7
safety==3.0.1
pre-commit==3.6.1
autopep8==2.0.4

# Type Stubs
types-redis==4.6.0.20240106
types-requests==2.31.0.20240125
sqlalchemy[mypy]==2.0.27

# Utilities
python-dotenv==1.0.1
click==8.1.7
typer==0.9.0
rich==13.7.0
tabulate==0.9.0
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2024.1
arrow==1.3.0
pendulum==3.0.0

# Email & Notifications
fastapi-mail==1.4.1
jinja2==3.1.3
sendgrid==6.11.0
twilio==8.13.0

# Workflow & State Machines
temporalio==1.5.1
prefect==2.16.1

# Geospatial
geopy==2.4.1
shapely==2.0.2

# Financial & Crypto
ccxt==4.2.25
stripe==7.14.0

# Monitoring SDKs
datadog==0.49.1
newrelic==9.5.0

# Additional Security
oauthlib==3.2.2
pynacl==1.5.0
secrets==1.0.2
